% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Analyz-class}
\alias{Analyz-class}
\title{Class Analyz}
\description{
Class to manage analysis steps described in a CSV file.
}
\section{Slots}{

\describe{
\item{\code{steps}}{A data frame attribute with the steps loaded from a CSV file.}

\item{\code{nrRows}}{A numeric attribute wiht the number of steps.}

\item{\code{nrColumns}}{A numeric attribute with the quantity of items of the largest step definition.}

\item{\code{stepItems}}{A list attribute with the items of a step.}

\item{\code{results}}{A list attribute with the execution result of the steps.}
}}

