\name{analyz-package}
\alias{analyz-package}
\alias{analyz}
\docType{package}
\title{
Data Analysis from CSV files
}
\description{
This package delivers methods to execute pre configured analysis steps over single data or loaded data matrices. The main purpose of this package is to provide a layer where a developer can implement on top of it an automation of data analysis contextualized on a purpose. The current version supports studies in biomonitoring, where a researcher colects data in the nature on a certain number of matrices and run different kinds of data analysis in order to get answers to her/his hipotesys.
}

\details{
\tabular{ll}{
Package: \tab analyz\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-10\cr
License: \tab GPL (>=2)\cr
Dependence: \tab methods\cr
}


}
\author{
Rodrigo Bühler: Who to complain to <rodrigo.buhler@ymail.com>

}

\keyword{ package }

\examples{

}
