% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.coerceType}
\alias{Analyz.coerceType}
\title{Method Analyz.coerceType}
\usage{
Analyz.coerceType(object, variable, type)
}
\arguments{
\item{object}{Description.}

\item{variable}{Description.}

\item{type}{Description.}
}
\value{
result  Description.
}
\description{
Method for variable type coercion.
}
\examples{
Analyz.coerceType()
}

