% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getNrColumns}
\alias{Analyz.getNrColumns}
\alias{Analyz.getNrColumns,Analyz,Analyz-method}
\alias{Analyz.getNrColumns,Analyz-method}
\title{Method Analyz.getNrColumns}
\usage{
Analyz.getNrColumns(object)

\S4method{Analyz.getNrColumns}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
nrColumns Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_nrCol <- Analyz.getNrColumns(obj)
}

