% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.getNrColumns}
\alias{Analyz.getNrColumns}
\title{Method Analyz.getNrColumns}
\usage{
Analyz.getNrColumns(object)
}
\arguments{
\item{object}{Description.}
}
\value{
nrColumns Description.
}
\description{
Description.
}
\examples{
Analyz.getNrColumns()
}

