% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.getNrLines}
\alias{Analyz.getNrLines}
\title{Method Analyz.getNrLines}
\usage{
Analyz.getNrLines(object)
}
\arguments{
\item{object}{Description.}
}
\value{
nrLines Description.
}
\description{
Description.
}
\examples{
Analyz.getNrLines()
}

