% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getNrRows}
\alias{Analyz.getNrRows}
\alias{Analyz.getNrRows,Analyz,Analyz-method}
\alias{Analyz.getNrRows,Analyz-method}
\title{Method Analyz.getNrRows}
\usage{
Analyz.getNrRows(object)

\S4method{Analyz.getNrRows}{Analyz}(object)
}
\arguments{
\item{object}{Object instance.}
}
\value{
nrRows Number of read rows from the analysis file.
}
\description{
Method for returning the number of rows read from the analysis file.
}
\examples{
obj <- new("Analyz")
v_NrRow <- Analyz.getNrRows(obj)
}

