% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Analyz.getParameters}
\alias{Analyz.getParameters}
\alias{Analyz.getParameters,Analyz,Analyz-method}
\alias{Analyz.getParameters,Analyz-method}
\title{Method Analyz.getParameters}
\usage{
Analyz.getParameters(object)

\S4method{Analyz.getParameters}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
stepParameters Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_paramters <- Analyz.getParameters(obj)
}

