% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getResult}
\alias{Analyz.getResult}
\alias{Analyz.getResult,Analyz,Analyz-method}
\alias{Analyz.getResult,Analyz-method}
\title{Method Analyz.getResult}
\usage{
Analyz.getResult(object, index)

\S4method{Analyz.getResult}{Analyz}(object, index)
}
\arguments{
\item{object}{Description.}

\item{index}{Description.}
}
\value{
result Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
vIdx <- numeric()
v_result <- Analyz.getResult(obj, vIdx)
}

