% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.getResult}
\alias{Analyz.getResult}
\title{Method Analyz.getResult}
\usage{
Analyz.getResult(object, index)
}
\arguments{
\item{object}{Description.}

\item{index}{Description.}
}
\value{
result Description.
}
\description{
Description.
}
\examples{
Analyz.getResult()
}

