% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Analyz.getStep}
\alias{Analyz.getStep}
\alias{Analyz.getStep,Analys,Analyz-method}
\alias{Analyz.getStep,Analyz-method}
\title{Method Analyz.getStep}
\usage{
Analyz.getStep(object, index)

\S4method{Analyz.getStep}{Analyz}(object, index)
}
\arguments{
\item{object}{Description.}

\item{index}{Description.}
}
\value{
stepList Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_step <- Analyz.getStep(obj, 1)
}

