% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.getStep}
\alias{Analyz.getStep}
\title{Method Analyz.getStep}
\usage{
Analyz.getStep(object, index)
}
\arguments{
\item{object}{Description.}

\item{index}{Description.}
}
\value{
stepList Description.
}
\description{
Description.
}
\examples{
Analyz.getStep()
}

