% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getStepCommand}
\alias{Analyz.getStepCommand}
\alias{Analyz.getStepCommand,Analyz,Analyz-method}
\alias{Analyz.getStepCommand,Analyz-method}
\title{Method Analyz.getStepCommand}
\usage{
Analyz.getStepCommand(object)

\S4method{Analyz.getStepCommand}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
stepCommand  Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_Command <- Analyz.getStepCommand(obj)
}

