% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Analyz.getStepItems}
\alias{Analyz,}
\alias{Analyz-method}
\alias{Analyz.getStepItems}
\alias{Analyz.getStepItems,}
\alias{Analyz.getStepItems,Analyz-method}
\title{Method Analyz.getStepItems}
\usage{
Analyz.getStepItems(object)

\S4method{Analyz.getStepItems}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
stepItems  Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_steps <- Analyz.getStepItems(obj)
}

