% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.getStepItems}
\alias{Analyz.getStepItems}
\title{Method Analyz.getStepItems}
\usage{
Analyz.getStepItems(object)
}
\arguments{
\item{object}{Description.}
}
\value{
line  Description.
}
\description{
Description.
}
\examples{
Analyz.getStepItems()
}

