% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getStepTitle}
\alias{Analyz.getStepTitle}
\alias{Analyz.getStepTitle,Analyz,Analyz-method}
\alias{Analyz.getStepTitle,Analyz-method}
\title{Method Analyz.getStepTitle}
\usage{
Analyz.getStepTitle(object)

\S4method{Analyz.getStepTitle}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
stepTitle  Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_Title <- Analyz.getStepTitle(obj)
}

