% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.loadSteps}
\alias{Analyz.loadSteps}
\alias{Analyz.loadSteps,Analyz,Analyz-method}
\alias{Analyz.loadSteps,Analyz-method}
\title{Method Analyz.loadSteps}
\usage{
Analyz.loadSteps(object, path)

\S4method{Analyz.loadSteps}{Analyz}(object, path)
}
\arguments{
\item{object}{Description.}

\item{path}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_path <- vector()
Analyz.loadSteps(obj, v_path)
}

