% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.loadSteps<-}
\alias{Analyz.loadSteps<-}
\title{Method Analyz.loadSteps}
\usage{
Analyz.loadSteps(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Analyz.loadSteps()
}

