% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.runAnalysis}
\alias{Analyz.runAnalysis}
\title{Method Analyz.runAnalysis}
\usage{
Analyz.runAnalysis(object, command, parameters)
}
\arguments{
\item{object}{Description.}

\item{command}{Description.}

\item{parameters}{Description.}
}
\value{
result    Description.
}
\description{
Description.
}
\examples{
Analyz.runAnalysis()
}

