% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.setResult<-}
\alias{Analyz.setResult<-}
\title{Method Analyz.setResult}
\usage{
Analyz.setResult(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Analyz.setResult()
}

