% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Analyz.setStepItems<-}
\alias{Analyz.setStepItems<-}
\title{Method Analyz.setStepItems}
\usage{
Analyz.setStepItems(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
Analyz.setStepItems()
}

