% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Global-class}
\alias{Global-class}
\title{Class Global}
\arguments{
\item{arg}{A list of arguments to be used in the class initializer.}
}
\description{
Description.
}
\section{Slots}{

\describe{
\item{\code{arguments}}{A list of arguments informed in time of creating the object.}

\item{\code{analysis}}{A character string with the path to the folder were the analysis CSV files are found.}

\item{\code{matrices}}{A character string with the path to the folder were the data matrices CSV files are found.}
}}

