% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{class}
\name{Analyz-class}
\alias{Analyz-class}
\title{Class to manage analysis steps described in a CSV file.}
\description{
Class Analyz
}
\section{Slots}{

\describe{
\item{\code{steps}}{A data frame attribute with the steps loaded from a CSV file.}

\item{\code{nrRows}}{A numeric attribute with the quantity of steps.}

\item{\code{nrColumns}}{A numeric attribute with the quantity of items of the largest step definition.}

\item{\code{stepItems}}{A list attribute with the items of a step.}

\item{\code{results}}{A list attribute with the execution result of the steps.}
}}

