% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.coerceType}
\alias{Analyz.coerceType}
\alias{Analyz.coerceType,Analyz,Analyz-method}
\alias{Analyz.coerceType,Analyz-method}
\title{Method Analyz.coerceType}
\usage{
Analyz.coerceType(object, variable, type)

\S4method{Analyz.coerceType}{Analyz}(object, variable, type)
}
\arguments{
\item{object}{Object instance.}

\item{variable}{Variable to be coerced.}

\item{type}{Data type to coerce the informed variable.}
}
\value{
result Coerced variable.
}
\description{
Method for variable type coercion.
}
\examples{
obj <- new("Analyz")
v_numeric <- Analyz.coerceType(obj, '2', 'numeric')
}

