% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getNrColumns}
\alias{Analyz.getNrColumns}
\alias{Analyz.getNrColumns,Analyz,Analyz-method}
\alias{Analyz.getNrColumns,Analyz-method}
\title{Method Analyz.getNrColumns}
\usage{
Analyz.getNrColumns(object)

\S4method{Analyz.getNrColumns}{Analyz}(object)
}
\arguments{
\item{object}{Object instance.}
}
\value{
nrColumns Number of read columns from the analysis file.
}
\description{
Method for returning the number of columns read from the analysis file.
}
\examples{
obj <- new("Analyz")
v_nrCol <- Analyz.getNrColumns(obj)
}

