% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.getNrRows}
\alias{Analyz.getNrRows}
\alias{Analyz.getNrRows,Analyz,Analyz-method}
\alias{Analyz.getNrRows,Analyz-method}
\title{Method Analyz.getNrRows}
\usage{
Analyz.getNrRows(object)

\S4method{Analyz.getNrRows}{Analyz}(object)
}
\arguments{
\item{object}{Description.}
}
\value{
nrRows Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_NrRow <- Analyz.getNrRows(obj)
}

