% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.loadSteps}
\alias{Analyz.loadSteps}
\alias{Analyz.loadSteps,Analyz,Analyz-method}
\alias{Analyz.loadSteps,Analyz-method}
\title{Method Analyz.loadSteps}
\usage{
Analyz.loadSteps(object, path)

\S4method{Analyz.loadSteps}{Analyz}(object, path)
}
\arguments{
\item{object}{Object instance.}

\item{path}{Path and analysis file name.}
}
\value{
object Object instance.
}
\description{
Method for reading the analysis file and fill the "steps" class spot.
}
\examples{
obj <- new("Analyz")
v_path <- vector()
Analyz.loadSteps(obj, v_path)
}

