% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Analyz.loadSteps<-}
\alias{Analyz.loadSteps,Analyz,Analyz-method}
\alias{Analyz.loadSteps<-}
\alias{Analyz.loadSteps<-,Analyz-method}
\title{Method Analyz.loadSteps<-}
\usage{
Analyz.loadSteps(object) <- value

\S4method{Analyz.loadSteps}{Analyz}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_path <- vector()
Analyz.loadSteps(obj) <- v_path
}

