% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.runAnalysis}
\alias{Analyz.runAnalysis}
\alias{Analyz.runAnalysis,Analyz,Analyz-method}
\alias{Analyz.runAnalysis,Analyz-method}
\title{Method Analyz.runAnalysis}
\usage{
Analyz.runAnalysis(object, command, parameters)

\S4method{Analyz.runAnalysis}{Analyz}(object, command, parameters)
}
\arguments{
\item{object}{Object instance.}

\item{command}{Description of the step command to be executed.}

\item{parameters}{Paramater(s) for the informed step command.}
}
\value{
result Result of the step execution.
}
\description{
Method for step execution.
}
\examples{
obj <- new("Analyz")
Analyz.runAnalysis(obj, 'mean', list(c(1,2,3)))
}

