% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.setResult<-}
\alias{Analyz.setResult,Analyz,Analyz-method}
\alias{Analyz.setResult<-}
\alias{Analyz.setResult<-,Analyz-method}
\title{Method Analyz.setResult<-}
\usage{
Analyz.setResult(object) <- value

\S4method{Analyz.setResult}{Analyz}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_result <- vector()
Analyz.setResult(obj) <- v_result
}

