% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.setResult<-}
\alias{Analyz.setResult,Analyz,Analyz-method}
\alias{Analyz.setResult<-}
\alias{Analyz.setResult<-,Analyz-method}
\title{Method Analyz.setResult<-}
\usage{
Analyz.setResult(object) <- value

\S4method{Analyz.setResult}{Analyz}(object) <- value
}
\arguments{
\item{object}{Object instance.}

\item{value}{Result value of an execution.}
}
\value{
object Object instance.
}
\description{
Method for storing the informed result to the "results" class spot.
}
\examples{
obj <- new("Analyz")
v_result <- vector()
Analyz.setResult(obj) <- v_result
}

