% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ClassAnalyz.R
\docType{methods}
\name{Analyz.setStepItems}
\alias{Analyz.setStepItems}
\alias{Analyz.setStepItems,Analyz,Analyz-method}
\alias{Analyz.setStepItems,Analyz-method}
\title{Method Analyz.setStepItems}
\usage{
Analyz.setStepItems(object, index)

\S4method{Analyz.setStepItems}{Analyz}(object, index)
}
\arguments{
\item{object}{Description.}

\item{index}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_index <- numeric()
Analyz.setStepItems(obj, v_index)
}

