% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Analyz.setStepItems<-}
\alias{Analyz.setStepItems,Analyz,Analyz-method}
\alias{Analyz.setStepItems<-}
\alias{Analyz.setStepItems<-,Analyz-method}
\title{Method Analyz.setStepItems<-}
\usage{
Analyz.setStepItems(object) <- value

\S4method{Analyz.setStepItems}{Analyz}(object) <- value
}
\arguments{
\item{object}{Description.}

\item{value}{Description.}
}
\value{
object Description.
}
\description{
Description.
}
\examples{
obj <- new("Analyz")
v_value <- vector()
Analyz.setStepItems(obj) <- v_value
}

